+++ playwright.config.js
@@ -18,7 +18,7 @@ module.exports = defineConfig({
     }
   ],
   webServer: {
-    command: 'pnpm build && npx http-server ./dist -p 4173 -c-1',
+    command: 'pnpm build && pnpm preview --host 127.0.0.1 --port 4173 --strictPort',
     url: 'http://127.0.0.1:4173',
     reuseExistingServer: !process.env.CI,
     timeout: 120 * 1000